restructure and initial markup and styling
