+++ src/app.ts
@@ -1,3 +1,3 @@
-import Controller from "./ts/controller";
+import Controller from "./ts/components/controller";
 
 const model = new Controller();
