+++ src/sass/_components.scss
@@ -1,0 +1,132 @@
+.task-container {
+  background-color: var(--color-primary-grey);
+  border-radius: 15px;
+
+  &__cards {
+    &::-webkit-scrollbar-track {
+      box-shadow: inset 0 0 0.6rem rgba(0, 0, 0, 0.3);
+      border-radius: 10px;
+      background-color: var(--color-white);
+    }
+
+    &::-webkit-scrollbar {
+      width: 0.4rem;
+      border-radius: 1rem;
+      background-color: var(--color-white);
+    }
+
+    &::-webkit-scrollbar-thumb {
+      border-radius: 1rem;
+      box-shadow: inset 0 0 0.6rem rgba(0, 0, 0, 0.3);
+      background-color: var(--color-primary);
+    }
+  }
+}
+
+.task-headline {
+  font-weight: 600;
+}
+
+.task-count {
+  color: var(--color-primary-dark);
+  font-size: 1.2rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 5px;
+}
+
+.btn-add {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+  height: 2rem;
+  border-radius: 7px;
+  background-color: var(--color-primary-light);
+  border: 1px solid var(--color-grey-light-4);
+  font-size: 1.75rem;
+  color: var(--color-primary-dark);
+  font-weight: 500;
+  padding-bottom: 8px;
+  transition: all 0.2s;
+
+  &:hover {
+    background-color: var(--color-primary);
+    color: var(--color-white);
+  }
+}
+
+.card-item {
+  padding: 1rem;
+  margin-right: 0.5rem;
+
+  background-color: var(--color-white);
+  border-radius: 15px;
+  border: 1px solid var(--color-grey-light-4);
+
+  transition: all 0.2s;
+
+  cursor: grab;
+
+  &__header {
+    font-weight: 500;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+  }
+
+  &__description {
+    color: var(--color-grey-dark-1);
+  }
+
+  &:hover {
+    border-color: var(--color-primary-dark);
+  }
+}
+
+.card-item__buttons {
+  & button {
+    background-color: unset;
+    outline: unset;
+    border: 0;
+    cursor: pointer;
+
+    & svg {
+      height: 1rem;
+      width: 1rem;
+      transition: all 0.2s;
+    }
+  }
+
+  & .btn-complete {
+    &:hover > svg {
+      fill: var(--color-primary-dark);
+    }
+
+    & svg {
+      fill: var(--color-primary);
+      height: 1.2rem;
+      width: 1.2rem;
+    }
+  }
+
+  & .btn-delete {
+    &:hover > svg {
+      fill: var(--color-red-active);
+    }
+    & svg {
+      fill: var(--color-red-inactive);
+      height: 1rem;
+      width: 1rem;
+    }
+  }
+}
+
+.heart {
+  color: red;
+}
+
+.link {
+  text-decoration: none;
+}
