+++ src/sass/_layout.scss
@@ -1,0 +1,72 @@
+body {
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
+  & > * {
+    width: 90%;
+  }
+}
+
+.header {
+  height: 10rem;
+  display: flex;
+  align-items: flex-end;
+  flex-shrink: 0;
+
+  &__title {
+    font-weight: 600;
+    font-size: 1.8rem;
+    padding-bottom: 1rem;
+  }
+}
+
+.container-grid {
+  flex: 1;
+  min-height: 0;
+  display: grid;
+  grid-template-rows: 1fr;
+  grid-template-columns: repeat(5, 1fr);
+  gap: 1rem;
+}
+
+.task-container {
+  display: flex;
+  flex-direction: column;
+  min-height: 0;
+  gap: 1rem;
+  padding: 0.75rem;
+  &__header {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+
+    & .container-info {
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
+
+      & > * {
+        height: 100%;
+      }
+    }
+  }
+  &__cards {
+    display: flex;
+    flex-direction: column;
+    gap: 0.75rem;
+
+    overflow-y: auto;
+    overflow-x: hidden;
+  }
+}
+
+.footer {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 7rem;
+  gap: 5px;
+}
