+++ src/ts/components/model.ts
